add go sample for ch01
